backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
uvicorn[standard]==0.32.0
pydantic==2.9.2
better-auth==0.0.15
asyncpg==0.30.0
alembic==1.13.3